Refactorisation du code du fichier bastion.py en bastion.py + connexion.py et SSHSErveur.py Limitation de l'utilisation de variable global
